+++ src/api/api_models/common/UnsignedTransactionMessage.ts
@@ -1,0 +1,3 @@
+export default interface UnsignedTransactionMessage {
+    unsignedTransaction: string;
+}