+++ src/components/account/list/UserRow.vue
@@ -1,8 +1,13 @@
 <template>
     <div
         :id="'user_' + user.username"
-        class="px-6 py-4 whitespace-no-wrap border-gray-200 cursor-pointer hover:bg-gray-200"
-        :class="{ 'rounded-t-lg': isFirstRow, 'rounded-b-lg': isLastRow, 'border-b': !isLastRow }"
+        class="px-6 py-4 whitespace-no-wrap border-gray-200"
+        :class="{
+            'rounded-t-lg': isFirstRow,
+            'rounded-b-lg': isLastRow,
+            'border-b': !isLastRow,
+            'cursor-pointer hover:bg-gray-200': user.isActive,
+        }"
         @click="editAccount(user.username)"
     >
         <div class="flex items-center">
@@ -10,7 +15,15 @@
                 <div class="flex items-center">
                     <img class="hidden sm:block w-12 h-12 rounded-full" :src="profilePicture" alt="profile_picture" />
                     <div class="sm:ml-4">
-                        <div class="text leading-5 font-medium text-blue-900 mb-1 truncate">{{ user.firstName }} {{ user.lastName }}</div>
+                        <div v-if="user.isActive" class="mb-1 truncate">
+                            <label class="text leading-5 font-medium text-blue-900 mr-2">{{ user.firstName }} {{ user.lastName }}</label>
+                        </div>
+                        <div v-if="!user.isActive" class="mb-1 truncate">
+                            <label v-if="isLecturer" class="text leading-5 font-medium text-blue-900 mr-2">
+                                {{ user.firstName }} {{ user.lastName }}
+                            </label>
+                            <label class="text-gray-600 italic">(inactive)</label>
+                        </div>
                         <div class="hidden sm:flex text leading-5 text-gray-500 truncate">@{{ user.username }}</div>
                         <span
                             class="sm:hidden inline-block text-xs px-2 rounded-lg font-semibold leading-5 tracking-wide mb-1 w-16 text-center"
@@ -39,7 +52,7 @@
                         </span>
                     </div>
 
-                    <div class="flex-col hidden sm:flex items-baseline" :class="[isStudent ? 'sm:flex' : 'sm:invisible']">
+                    <div class="flex-col hidden sm:flex items-baseline" :class="[isStudent && user.isActive ? 'sm:flex' : 'sm:invisible']">
                         <div class="leading-5 text-blue-900 ml-1 mb-1">{{ student.matriculationId }}</div>
                         <div class="hidden sm:flex items-center leading-5 text-gray-500">
                             <span class="mr-2 fa-stack text-xs" style="font-size: 0.63em">
@@ -86,13 +99,12 @@
         setup(props: any) {
             let profilePicture = ref("");
             function editAccount(username: string) {
-                router.push({ path: "/editAccount/" + username });
+                if (props.user.isActive) router.push({ path: "/editAccount/" + username });
             }
             const isStudent = props.user.role === Role.STUDENT;
             const isLecturer = props.user.role === Role.LECTURER;
             const isAdmin = props.user.role === Role.ADMIN;
             const student = props.user as Student;
-
             profilePicture.value = process.env.VUE_APP_API_BASE_URL + "/user-management/users/" + props.user.username + "/thumbnail?";
 
             return { editAccount, isStudent, isLecturer, isAdmin, student, profilePicture };