+++ tests/e2e/helpers/NavigationHelper.ts
@@ -6,6 +6,10 @@ export function navigateToCourseListAdmin() {
     cy.visit("all-courses");
 }
 
+export function navigateToExamRegForm() {
+    cy.visit("create-exam-reg");
+}
+
 export function navigateToAccountForm() {
     cy.visit("createAccount");
 }
@@ -40,4 +44,8 @@ export function navigateToPrivateProfile() {
 
 export function navigateToAboutPage() {
     cy.visit("about");
+}
+
+export function navigateToImmatriculationPage() {
+    cy.visit("immatriculation");
 }