+++ tests/e2e/navbarDesktop.spec.ts
@@ -49,6 +49,13 @@ describe("admin", () => {
         cy.get("#nav_desktop_admin_menu_courses_create").click();
         cy.url().should("contain", "createCourse");
     });
+
+    it("exam-regs", () => {
+        // create exam reg form
+        cy.get("#nav_desktop_admin_menu_exreg").trigger("mouseover");
+        cy.get("#nav_desktop_admin_menu_exreg_create").click();
+        cy.url().should("contain", "create-exam-reg");
+    });
     checkProfile();
 });
 
@@ -89,6 +96,16 @@ describe("student", () => {
         cy.get("#nav_desktop_student_menu_courses_all").click();
         cy.url().should("contain", "courses");
     });
+
+    it("study", () => {
+        //open the menu
+        cy.get("#nav_desktop_student_menu_study").trigger("mouseover");
+
+        //all courses
+        cy.get("#nav_desktop_student_menu_study_immatriculation").click();
+        cy.url().should("contain", "immatriculation");
+    });
+
     checkProfile();
 });