+++ CHANGELOG.md
@@ -1,7 +1,33 @@
-# [v0.13.1 WIP](https://github.com/upb-uc4/ui-web/compare/v0.13.0...v0.13.1) (2020-XX-XX)
+# [WIP 0.14.1](https://github.com/upb-uc4/ui-web/compare/v0.14.0...v0.14.1) (2020-12-XX)
+## Feature
+- add hyperledger versions [#754](https://github.com/upb-uc4/ui-web/pull/754)
+- add frontened signing for matriculation use case [#749](https://github.com/upb-uc4/ui-web/pull/749)
+
+# [v0.14.0](https://github.com/upb-uc4/ui-web/compare/v0.13.1...v0.14.0) (2020-12-08)
+## Feature
+- matriculation using examination regulations [#745](https://github.com/upb-uc4/ui-web/pull/745)
+- Add create exam regulation form [#690](https://github.com/upb-uc4/ui-web/pull/690)
+
+## Bugfix
+- refactor course list items [#715](https://github.com/upb-uc4/ui-web/pull/715)
+
+# [v0.13.1-hotfix.1](https://github.com/upb-uc4/ui-web/compare/v0.13.1...v0.13.1-hotfix.1) (2020-11-27)
+## Bugfix
+- hide create course button in course catalogue [#747](https://github.com/upb-uc4/ui-web/pull/747)
+
+# [v0.13.1](https://github.com/upb-uc4/ui-web/compare/v0.13.0...v0.13.1) (2020-11-26)
+
+## Feature
+- added governmentId to user creation [#701](https://github.com/upb-uc4/ui-web/pull/701)
+- add soft deletion of users [#702](https://github.com/upb-uc4/ui-web/pull/702), [#704](https://github.com/upb-uc4/ui-web/pull/704)
+
+## Refactoring
+- Merge endpoints for user creation [#700](https://github.com/upb-uc4/ui-web/pull/700)
+- Remove unused components [#739](https://github.com/upb-uc4/ui-web/pull/739)
 
 ## Bugfix
 - add ECTS field to course form [#738](https://github.com/upb-uc4/ui-web/pull/738)
+- reset store state on logout [#732](https://github.com/upb-uc4/ui-web/pull/732)
 
 # [v0.13.0](https://github.com/upb-uc4/ui-web/compare/v0.12.0...v0.13.0) (2020-11-23)
 
@@ -15,6 +41,7 @@
 - add a message for the student if not immatriculated to any semester yet [#692](https://github.com/upb-uc4/ui-web/pull/692)
 - Replace deprecated suspense wrapper by onBeforeMount hooks and a "busy" variable [#714](https://github.com/upb-uc4/ui-web/pull/714)
 - restyle login page [#698](https://github.com/upb-uc4/ui-web/issues/698)
+
 
 ## Bugfix
 - fix mobile view of login page [#698](https://github.com/upb-uc4/ui-web/issues/698)